requirements copy.txt
autopep8
distinctipy
einops==0.3.0
gdown
h5py==3.1.0
ipython
joblib>=1.0.1
jupyterlab
kornia== 0.6.11
loguru==0.5.3
matplotlib
numpy
omegaconf
# onnx
# onnxruntime-gpu
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
pydegensac
pylint
pytorch-lightning==1.3.5
rich
scikit-image
scipy
tables
tabulate==0.8.10
termcolor  
torchinfo
torchmetrics==0.6.0
cs
tqdm
yacs>=0.1.8